No more flat directory structure.
Originally, the intention was to have HTML templates in the root
of the repository so it's easy to contribute for any person.
However, this led to quite a mess, so I'm moving files and
templates into their own folders and let's rely on direct
links to HTML files on GitHub in the footer of the site.

I've chosen names like `pages`, `files` and `web.py`, so it's
beginner friendly, but I'm not sure whether it won't be
misleading more than helping. Let's see, it's easy to change.